% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilsimage.R
\name{changepngtocormorantstyle}
\alias{changepngtocormorantstyle}
\title{.}
\usage{
changepngtocormorantstyle(image, backgroundcolor = 98L)
}
\arguments{
\item{image}{.}

\item{backgroundcolor}{.}
}
\value{
.
}
\description{
.. content for  (no empty lines) ..
}
\details{
.. content for  ..
}
\examples{
1
}
\author{
emilio
}
