% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hashmap.R
\name{convert_key_to_vector}
\alias{convert_key_to_vector}
\title{.}
\usage{
convert_key_to_vector(x)
}
\arguments{
\item{x}{A vector of integers. It reprents a key. See convert_vector_to_key.}
}
\value{
.
}
\description{
.. content for  (no empty lines) ..
}
\details{
.. content for  ..
}
\examples{
## See convert_vector_to_key
}
\author{
emilio
}
