% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hashmap.R
\name{convert_vector_to_key}
\alias{convert_vector_to_key}
\title{Convert a vector of integers, numeric, characters into a vector of integers}
\usage{
convert_vector_to_key(x)
}
\arguments{
\item{x}{A vector of logical, integers, numeric of characters}
}
\value{
A vector with type (of x), size of integers, and n=size integers with the representation of x
}
\description{
.. content for  (no empty lines) ..
}
\details{
It returns the internal represenatation of the keys and values of the hashmap.
Useful if you only want to deal with integers.
}
\examples{
a <- 1:2
k <- convert_vector_to_key(a)
b <- convert_key_to_vector(k)
identical(a,b)

a <- 1.*c(1:2)
k <- convert_vector_to_key(a)
b <- convert_key_to_vector(k)
identical(a,b)

a <- as.character(c(1:2))
k <- convert_vector_to_key(a)
b <- convert_key_to_vector(k)
identical(a,b)
}
\author{
emilio
}
