% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilsimage.R
\name{convertfilepdftopng}
\alias{convertfilepdftopng}
\title{.}
\usage{
convertfilepdftopng(
  infilepdf,
  outfilepng = "slides/slide",
  width = 1920,
  height = 1080
)
}
\arguments{
\item{infilepdf}{.}

\item{outfilepng}{.}

\item{width}{.}

\item{height}{.}
}
\value{
.
}
\description{
.. content for  (no empty lines) ..
}
\details{
.. content for  ..
}
\examples{
## unlink("slides",recursive=TRUE)
## dir.create("slides")
## infilepdf <- "164_presentacion.pdf"
## convertfilepdftopng(infilepdf,outfilepng="slides/slide", width = 1920, height = 1080)
}
\author{
emilio
}
