% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crearvideo.R
\name{crearframesyvideo}
\alias{crearframesyvideo}
\title{.}
\usage{
crearframesyvideo(
  img,
  m,
  slide,
  nseconds,
  directoriodiapos = "slides",
  colortexto = c(898898898, 890890890),
  colorfondo = 98098098,
  precision = 1000L,
  framerate = 25L
)
}
\arguments{
\item{img}{image}

\item{m}{matrix with detailed information}

\item{slide}{number of slide}

\item{nseconds}{nseconds}

\item{directoriodiapos}{.}

\item{colortexto}{.}

\item{colorfondo}{.}

\item{precision}{.}

\item{framerate}{.}
}
\value{
.
}
\description{
.. content for  (no empty lines) ..
}
\details{
.. content for  ..
}
\author{
emilio
}
