% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crearvideo.R
\name{crearvideo}
\alias{crearvideo}
\title{.}
\usage{
crearvideo(img, slide, nseconds, writePNG, ...)
}
\arguments{
\item{img}{image}

\item{slide}{number of image}

\item{nseconds}{number of seconds}

\item{writePNG}{png::writePNG}

\item{...}{.}
}
\value{
.
}
\description{
.. content for  (no empty lines) ..
}
\details{
.. content for  ..
}
\author{
emilio
}
