% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crearvideo.R
\name{crearvideofake}
\alias{crearvideofake}
\title{.}
\usage{
crearvideofake(img, slide, nseconds = 5)
}
\arguments{
\item{img}{image}

\item{slide}{slide}

\item{nseconds}{nseconds}
}
\value{
.
}
\description{
.. content for  (no empty lines) ..
}
\details{
.. content for  ..
}
\author{
emilio
}
