% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hashtable.R
\name{create_ht}
\alias{create_ht}
\title{Create a hash set of a vector or a matrix.}
\usage{
create_ht(x, first = TRUE, cols = NULL)
}
\arguments{
\item{x}{A vector or a matrix of integers, numeric or characters}

\item{first}{If you want to search for the first ocurrence (TRUE) or the last occurrence (FALSE)}

\item{cols}{It x is a matrix, you can select the columns to process}
}
\value{
A hash table.
}
\description{
.. content for  (no empty lines) ..
}
\details{
In case of a matrix, it indexes each row.
}
\examples{
ht <- create_ht(1:10)
}
\author{
emilio
}
