% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hashmap.R
\name{create_map}
\alias{create_map}
\title{.}
\usage{
create_map(size)
}
\arguments{
\item{size}{Tentative size. Not very important}
}
\value{
A pointer of HashMap
}
\description{
Create a HashMap of keys and values
}
\details{
.. content for  ..
}
\examples{
k <- list(0L, 1., "2", c(0L,0L),c(1,1), c("2","2"))
v <- list(10L, 11., "20", c(10L,10L),c(11,11), c("20","10"))
map <- create_map(2L)
for(key in k){
    for(value in v){
        insert_map(key,value,map)
    }
}
for(key in k){
    cat("\nKey:\n")
    print(key)
    cat("Value:\n")
    print(get_map(key,map))
}
 
size  <-  length(map)
for(i in 1L:size) {
    cat("\nPair Key Value:\n")
    print(get_key_by_position_map(i,map))
    print(get_value_by_position_map(i,map))
 }
 
 insert_map(0L,"CCCC",map)
 get_map(0L,map)

size <- 10 # try 10^6
t <- 1:size
pos <- as.character(t + (t-1)*size)
map <- create_map(size)
a <- tic_named("Brute force")
for(i in 1:length(pos)) insert_map(pos[i],pos[i], map)
toc_print_timer(a)
map2 <- create_map(size)
 k <- as.list(pos)
v <- as.list(pos)
b <- tic_named("Using lists. It is faster if you have the lists")
insert_map(k,v, map2)
toc_print_timer(b)
}
\author{
emilio
}
