% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparsematrix.R
\name{create_sm}
\alias{create_sm}
\title{Create a sparse matrix using hash tables}
\usage{
create_sm(nrows, ncols)
}
\arguments{
\item{nrows}{nrows}

\item{ncols}{ncols}
}
\value{
.
}
\description{
.. content for  (no empty lines) ..
}
\details{
.. content for  ..
}
\examples{
sm <- create_sm(10,10)
}
\author{
emilio
}
