% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparsematrix.R
\name{dim.etm_SparseMatrix}
\alias{dim.etm_SparseMatrix}
\title{.}
\usage{
\method{dim}{etm_SparseMatrix}(x)
}
\arguments{
\item{x}{sm}
}
\value{
dimension of sm
}
\description{
.. content for  (no empty lines) ..
}
\details{
.. content for  ..
}
\examples{
sm <-  create_sm(10,7)
dim(sm)
}
\author{
emilio
}
