% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onLoad.R
\name{.onUnload}
\alias{.onUnload}
\title{Always include this file}
\usage{
.onUnload(libpath)
}
\arguments{
\item{libpath}{1}
}
\value{
1
}
\description{
i) To export the library.
   Something similar to @rawNamespace useDynLib(matchmatrix); useDynLib(emiliocpp)
}
\details{
ii) import methods is included to skip the warning:
* checking dependencies in R code ... NOTE
Namespace in Imports field not imported from: ‘methods’
All declared Imports should be used.
}
\author{
emilio
}
