% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hashtable.R
\name{dump_ht}
\alias{dump_ht}
\title{Returns the indexes stored in the hash table.}
\usage{
dump_ht(x)
}
\arguments{
\item{x}{A hash table}
}
\value{
a vector of integers.
}
\description{
.. content for  (no empty lines) ..
}
\details{
.. content for  ..
}
\examples{
x <- c(1:10,10:1)
ht <- create_ht(x)
dump_ht(ht)
}
\author{
emilio
}
