% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparsematrix.R
\name{dump_sm}
\alias{dump_sm}
\title{Return all the values}
\usage{
dump_sm(sm)
}
\arguments{
\item{sm}{sm}
}
\value{
A matrix (i,j,values)
}
\description{
.. content for  (no empty lines) ..
}
\details{
.. content for  ..
}
\examples{
sm <- create_sm(10,10)
insert_sm(1:10,1:10,10:1,sm)
dump_sm(sm)
}
\author{
emilio
}
