% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hashtable.R
\name{find_ht}
\alias{find_ht}
\title{Get the first position of value in x.}
\usage{
find_ht(value, x, ht, notfound = NA_integer_, cols = NULL)
}
\arguments{
\item{value}{Is value in x? Return the position of value in x.}

\item{x}{A vector or a matrix of integers, numeric or characters}

\item{ht}{the hash table created with x and cols}

\item{notfound}{An integer value}

\item{cols}{It x is a matrix, you can select the columns to process}
}
\value{
A vector of integers with the position of value in x.
}
\description{
It works as match.
}
\details{
.. content for  ..
}
\examples{
x <- 1:10
ht <- create_ht(x)
find_ht(c(1L,30L),x,ht)
}
\author{
emilio
}
