% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtoslides.R
\name{generatevideoslide}
\alias{generatevideoslide}
\title{Convert a slide of beamer cormorant style to a video, writting each symbol in a frame.}
\usage{
generatevideoslide(
  slide,
  nseconds,
  readPNG,
  writePNG,
  npixelslettermin = 15L,
  directoriodiapo = "slides"
)
}
\arguments{
\item{slide}{number of the slide to convert to video.}

\item{nseconds}{Duration of the video}

\item{readPNG}{Use this function png::readPNG}

\item{writePNG}{Use this function png::writePNG}

\item{npixelslettermin}{minimum number of pixels for letters}

\item{directoriodiapo}{directory of diapo}
}
\value{
Save a video in the hard disk.
}
\description{
Slides must be in the subdirectory 'slides', and must be of the form 'slide-03.png'.
}
\details{
We need the external program 'ffmpeg'.
}
\examples{
1
}
\author{
emilio
}
