% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msplit.R
\name{get_first_element_of_list_of_integers}
\alias{get_first_element_of_list_of_integers}
\title{Get the first element of a list of vector of integers.}
\usage{
get_first_element_of_list_of_integers(listofintegers)
}
\arguments{
\item{listofintegers}{A list of vector of integers}
}
\value{
A vector of integers.
}
\description{
Get the first element of a list of integers.
}
\examples{
x <- list(c(1L,2L),c(3L,4L,5L,100L,200L),c(6L,7L))
get_first_element_of_list_of_integers(x)
}
\author{
emilio
}
