% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hashmap.R
\name{get_map}
\alias{get_map}
\title{.}
\usage{
get_map(key, map)
}
\arguments{
\item{key}{Vector of logical, integer, numeric or characters}

\item{map}{A HashTable}
}
\value{
value stored with that key. NULL if not exists that key
}
\description{
.. content for  (no empty lines) ..
}
\details{
.. content for  ..
}
\examples{
## see create_map
}
\author{
emilio
}
