% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparsematrix.R
\name{get_sm}
\alias{get_sm}
\title{.}
\usage{
get_sm(rows, cols, sm, zerovalue = 0L)
}
\arguments{
\item{rows}{rows}

\item{cols}{cols}

\item{sm}{sparse matrix}

\item{zerovalue}{If the cells are empty, return this value.}
}
\value{
If not empty, return the value of teh cells
}
\description{
.. content for  (no empty lines) ..
}
\details{
.. content for  ..
}
\examples{
sm <- create_sm(10,10)
insert_sm(1,1,7,sm)
get_sm(1,1,sm)
get_sm(2,2,sm)
}
\author{
emilio
}
