% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hashmap.R
\name{get_value_by_position_map}
\alias{get_value_by_position_map}
\title{.}
\usage{
get_value_by_position_map(pos, map)
}
\arguments{
\item{pos}{1,.., length(map)}

\item{map}{HashMap}
}
\value{
The value of that position
}
\description{
.. content for  (no empty lines) ..
}
\details{
.. content for  ..
}
\examples{
## see create_map
}
\author{
emilio
}
