% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getnumberprocessors.R
\name{getnumberprocessors}
\alias{getnumberprocessors}
\title{Returns the number of processors available to the device.}
\usage{
getnumberprocessors()
}
\value{
The number of processors available.
}
\description{
Returns the number of processors available to the device.
}
\details{
This function returns the number of processors available to the device.
}
\examples{
getnumberprocessors()
}
\author{
Emilio
}
