% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hashmap.R
\name{insert_map}
\alias{insert_map}
\title{.}
\usage{
insert_map(key, value, map)
}
\arguments{
\item{key}{(A list of) Vector of logical, integer, numeric or characters}

\item{value}{(A list of) Vector of logical, integer, numeric or characters}

\item{map}{A HashTable}
}
\value{
An invisible 0( fails)/1(success)
}
\description{
.. content for  (no empty lines) ..
}
\details{
With list of vector is faster. Otherwise you can insert element by element
}
\examples{
## see create_map
}
\author{
emilio
}
