% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hashtable.R
\name{isin}
\alias{isin}
\title{Similar to in}
\usage{
isin(x, table, cols = NULL)
}
\arguments{
\item{x}{x}

\item{table}{table}

\item{cols}{cols}
}
\value{
.
}
\description{
.. content for  (no empty lines) ..
}
\details{
.. content for  ..
}
\examples{
isin(c(1,20),c(1,10))
}
\author{
emilio
}
