% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hashmap.R
\name{length.etm_HashMap}
\alias{length.etm_HashMap}
\title{.}
\usage{
\method{length}{etm_HashMap}(x)
}
\arguments{
\item{x}{HashMap}
}
\value{
.
}
\description{
.. content for  (no empty lines) ..
}
\details{
.. content for  ..
}
\examples{
## see create_map
}
\author{
emilio
}
