% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparsematrix.R
\name{length.etm_SparseMatrix}
\alias{length.etm_SparseMatrix}
\title{.}
\usage{
\method{length}{etm_SparseMatrix}(x)
}
\arguments{
\item{x}{x}
}
\value{
number of cells with information.
}
\description{
.. content for  (no empty lines) ..
}
\details{
.. content for  ..
}
\author{
emilio
}
