% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdoby.R
\name{mdoby}
\alias{mdoby}
\title{Do something FUN to each group of rows splitted by some columns in a matrix}
\usage{
mdoby(x, FUN, ..., by, rows = NULL)
}
\arguments{
\item{x}{A matrix of integer, double or character data.}

\item{FUN}{A function to apply to a subset of the matrix. Should return vectors of the same type of x (integer, double or character).}

\item{...}{Additional arguments to FUN.}

\item{by}{Split the matrix by the values of these columns.}

\item{rows}{A logical filter or integer vector selecting the rows of the matrix. By default it uses all the rows.}
}
\value{
A matrix with the values of the columns used in by and the results of the FUN.
}
\description{
A tiny version of data.table[rows,FUN,by] focused in matrices.
}
\details{
The results of the FUN are silently coerced to the type of the matrix (integer,double or character).
}
\examples{
x <- as.matrix(iris[,-5])
f <- iris[["Species"]] == "virginica"
mdoby(x,function(x){c(n=nrow(x),min=min(x[,"Sepal.Length"]),max=max(x[,"Sepal.Length"]))}, by=c("Petal.Length"),rows=f)
}
\author{
emilio
}
