% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmerge.R
\name{mmerge}
\alias{mmerge}
\title{Merge two matrices.}
\usage{
mmerge(
  x,
  y,
  by = NULL,
  by.x = by,
  by.y = by,
  rowsx = NULL,
  rowsy = NULL,
  suffixes = c(".x", ".y")
)
}
\arguments{
\item{x}{The first matrix.}

\item{y}{The second matrix.}

\item{by}{Columns used for merging.}

\item{by.x}{Columns from matrix x used for merging.}

\item{by.y}{Columns from matrix y used for merging.}

\item{rowsx}{Rows from matrix x to be included in the result.
Can be either an integer vector or a logical vector.}

\item{rowsy}{Rows from matrix y to be included in the result.
Can be either an integer vector or a logical vector.}

\item{suffixes}{Suffixes to append to column names when they are duplicated.}
}
\value{
A merged matrix with all the rowsx of x.
}
\description{
Merge two matrices.
}
\details{
This function merges two matrices, x and y, based on specified columns and rows.
All rows of x are selected. If there are duplicated values on y, it only takes the first occurrence.
If there is no match, the row is filled with NA.
}
\examples{
x  <- matrix(sample(1:10,1e3,TRUE),ncol=2)
y  <- matrix(1:10,ncol=5)
colnames(x)  <- c("v1","v2")
colnames(y) <- c("v1","v2","v3","v4","v5")
res <- mmerge(x,y)
str(res)
}
\author{
emilio
}
