% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrescale.R
\name{mrescale}
\alias{mrescale}
\title{Rescale a matrix}
\usage{
mrescale(m, nrow, ncol)
}
\arguments{
\item{m}{a numeric matrix}

\item{nrow}{new dimension}

\item{ncol}{new dimension}
}
\value{
a matrix with dimension nrow x ncol (numeric)
}
\description{
Rescale a matrix using moving averages in X and Y.
}
\details{
Useful fo images.
}
\examples{
m <- matrix(1:6,2,3)
mrescale(m,4,6)
}
\author{
emilio
}
