% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msegments.R
\name{msegments}
\alias{msegments}
\title{.}
\usage{
msegments(x, dims = dim(x))
}
\arguments{
\item{x}{points of a matrix}

\item{dims}{dimension of the matrix}
}
\value{
.
}
\description{
.. content for  (no empty lines) ..
}
\details{
.. content for  ..
}
\examples{
x <- as.integer(c(2,3,5,6,7,9))
msegments(x,dims=c(10L,1L))
}
\author{
emilio
}
