% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msplit.R
\name{msplit}
\alias{msplit}
\title{Divide a matrix into groups defined by the columns 'cols'.}
\usage{
msplit(x, by = NULL, rows = NULL)
}
\arguments{
\item{x}{A matrix.}

\item{by}{A character or integer vector specifying the columns for grouping. NULL means all the columns.}

\item{rows}{An integer or logical vector indicating the rows to include. NULL means all the rows.}
}
\value{
A list of integers. Each element of this list is a vector with the positions of the rows of each group.
}
\description{
Get a list of rows for each group defined by columns.
}
\details{
This function divides a matrix into groups defined by the specified columns.
}
\examples{
x <- as.matrix(iris)
a <- msplit(x,by="Species")
str(a)
}
\author{
emilio
}
