% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcbind.R
\name{msubset}
\alias{msubset}
\title{Extract a submatrix.}
\usage{
msubset(x, rows = NULL, cols = NULL)
}
\arguments{
\item{x}{The input matrix.}

\item{rows}{Rows to extract. Can be either an integer vector or a logical vector.}

\item{cols}{Columns to extract. Can be either an integer vector or a character vector.}
}
\value{
The extracted submatrix.
}
\description{
Extract a submatrix using threads.
}
\details{
This function uses threads to efficiently extract a submatrix from the input matrix x.
}
\examples{
x <- as.matrix(iris)
res1 <- x[1:10,3:4,drop=FALSE]
res2 <- msubset(x,1:10,3:4)
all.equal(res1,res2)
}
\author{
emilio
}
