% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/munique.R
\name{munique}
\alias{munique}
\title{A fast alternative for 'unique' with matrices.}
\usage{
munique(x, cols = NULL, rows = NULL)
}
\arguments{
\item{x}{The input matrix.}

\item{cols}{Columns to consider for uniqueness.}

\item{rows}{A logical or integer vector. If logical, its length must be nrow(x).}
}
\value{
A matrix with no duplicated rows.
}
\description{
A fast alternative for 'unique' with matrices.
}
\details{
This function provides a fast alternative for 'unique' when dealing with matrices.
}
\examples{
x <- matrix(rep(1:10,4),ncol=2)
all.equal(munique(x),unique(x))
}
\author{
emilio
}
