% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obtenerboundingbox.R
\name{obtenerboundingbox}
\alias{obtenerboundingbox}
\title{Obtain the bounding boxes}
\usage{
obtenerboundingbox(img)
}
\arguments{
\item{img}{An image}
}
\value{
A matrix with the bounding boxes.
}
\description{
.. content for  (no empty lines) ..
}
\details{
.. content for  ..
}
\examples{
1
}
\author{
emilio
}
