% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obtenerboundingbox.R
\name{obtenersegmentos}
\alias{obtenersegmentos}
\title{Pixels y segmentos de la imagen}
\usage{
obtenersegmentos(img)
}
\arguments{
\item{img}{imagen}
}
\value{
.
}
\description{
.. content for  (no empty lines) ..
}
\details{
.. content for  ..
}
\examples{
1
}
\author{
emilio
}
