% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crearvideo.R
\name{ordenarpixels}
\alias{ordenarpixels}
\title{ordena los pixels de la imagen}
\usage{
ordenarpixels(img)
}
\arguments{
\item{img}{imagen}
}
\value{
.
}
\description{
.. content for  (no empty lines) ..
}
\details{
.. content for  ..
}
\examples{
1
}
\author{
emilio
}
