% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_array.R
\name{plot_array}
\alias{plot_array}
\title{Plot an array as image.}
\usage{
plot_array(
  image,
  rowlimits = c(1L, dim(image)[1]),
  collimits = c(1L, dim(image)[2]),
  ...
)
}
\arguments{
\item{image}{A matrix or array.}

\item{rowlimits}{To zoom the image.}

\item{collimits}{To zoom the image.}

\item{...}{.}
}
\value{
.
}
\description{
.. content for  (no empty lines) ..
}
\details{
.. content for  ..
}
\examples{
plot_array(matrix(0:1, ncol = 50, nrow = 27))
}
\author{
emilio
}
