% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparsematrix.R
\name{print.etm_SparseMatrix}
\alias{print.etm_SparseMatrix}
\title{.}
\usage{
\method{print}{etm_SparseMatrix}(x, ...)
}
\arguments{
\item{x}{sm}

\item{...}{dots}
}
\value{
.
}
\description{
.. content for  (no empty lines) ..
}
\details{
.. content for  ..
}
\author{
emilio
}
