% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparsematrix.R
\name{print_sm}
\alias{print_sm}
\title{.}
\usage{
print_sm(sm, n = 6L, verbose = TRUE)
}
\arguments{
\item{sm}{sparse matrix}

\item{n}{n rows to print. 0 to print all.}

\item{verbose}{verbose}
}
\value{
.
}
\description{
.. content for  (no empty lines) ..
}
\details{
.. content for  ..
}
\author{
emilio
}
