% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processcols.R
\name{processcols}
\alias{processcols}
\title{Transform into an integer vector the specified columns.}
\usage{
processcols(x, cols = NULL)
}
\arguments{
\item{x}{An object with columns.}

\item{cols}{A character or an integer vector specifying the names or positions of the columns.}
}
\value{
An integer vector with the position of that column in x.
}
\description{
Transform specified columns into an integer vector.
}
\details{
This function transforms specified columns of an object into an integer vector.
If all the values are negative, then it selects the other columns.
If a name does not appear in the names of the columns, it is removed.
}
\examples{
processcols(iris,c("Species")) # 5
processcols(iris,-c(2:4))
}
\author{
emilio
}
