% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table2.R
\name{table2}
\alias{table2}
\title{Univariate table}
\usage{
table2(x)
}
\arguments{
\item{x}{An atomic vector}
}
\value{
A list with the frequencies (lenths) and the values, ordered by frequency.
}
\description{
.. content for  (no empty lines) ..
}
\details{
..
}
\examples{
table2(c(1,2,2,2,3,3))
}
\author{
emilio
}
