% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tictoc.R
\name{tic_named}
\alias{tic_named}
\title{.}
\usage{
tic_named(x)
}
\arguments{
\item{x}{A character}
}
\value{
a integer number
}
\description{
.. content for  (no empty lines) ..
}
\details{
.. content for  ..
}
\examples{
a <- tic_named("Hello")
toc_print_timer(a)
}
\author{
emilio
}
