% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tictoc.R
\name{toc_timer}
\alias{toc_timer}
\title{.}
\usage{
toc_timer(x)
}
\arguments{
\item{x}{The integer returned by  tic_timer}
}
\value{
.
}
\description{
.. content for  (no empty lines) ..
}
\details{
.. content for  ..
}
\examples{
a <- tic_named("HH")
toc_timer(a)
}
\author{
emilio
}
